+++ examples/mcp/graph.rs
@@ -17,37 +17,39 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-use ddo::core::common::Variable;
-use regex::Regex;
 use std::fs::File;
 use std::io::{BufRead, BufReader, Lines, Read};
 use std::ops::{Index, IndexMut};
+
+use regex::Regex;
+
+use ddo::common::Variable;
 
 /// The graph is represented by its adjacency matrix
 #[derive(Debug, Clone)]
 pub struct Graph {
     pub nb_vertices: usize,
-    pub adj_matrix : Vec<i32>
+    pub adj_matrix : Vec<isize>
 }
 
 impl Graph {
     pub fn new(n: usize) -> Self {
         Graph {nb_vertices: n, adj_matrix: vec![0; n * n]}
     }
-    pub fn sum_of_negative_edges(&self) -> i32 {
+    pub fn sum_of_negative_edges(&self) -> isize {
         // sum is divided by two because the graph should be symmetrical
         self.adj_matrix.iter().cloned()
             .filter(|x| x.is_negative())
-            .sum::<i32>() / 2
+            .sum::<isize>() / 2
     }
-    pub fn add_bidir_edge(&mut self, x: usize, y: usize, w: i32) {
+    pub fn add_bidir_edge(&mut self, x: usize, y: usize, w: isize) {
         self[(x, y)] = w;
         self[(y, x)] = w;
     }
 
     pub fn from_lines<B: BufRead>(lines: Lines<B>) -> Graph {
-        let graph = Regex::new(r"^(?P<vars>\d+)\s+(?P<edges>\d+)$").unwrap();
-        let edge  = Regex::new(r"^(?P<src>\d+)\s+(?P<dst>\d+)\s+(?P<w>-?\d+)$").unwrap();
+        let graph   = Regex::new(r"^(?P<vars>\d+)\s+(?P<edges>\d+)$").unwrap();
+        let edge    = Regex::new(r"^(?P<src>\d+)\s+(?P<dst>\d+)\s+(?P<w>-?\d+)$").unwrap();
 
         let mut result = Graph::new(0);
         for line in lines {
@@ -57,6 +59,7 @@ impl Graph {
             if line.is_empty() {
                 continue;
             }
+            // skip comments
             if line.starts_with("c ") {
                 continue;
             }
@@ -69,7 +72,7 @@ impl Graph {
             if let Some(caps) = edge.captures(&line) {
                 let x = caps["src"].to_string().parse::<usize>().unwrap() - 1;
                 let y = caps["dst"].to_string().parse::<usize>().unwrap() - 1;
-                let w = caps["w"].to_string().parse::<i32>().unwrap();
+                let w = caps["w"].to_string().parse::<isize>().unwrap();
                 result.add_bidir_edge(x, y, w);
             }
         }
@@ -81,25 +84,25 @@ impl Graph {
     }
 }
 impl Index<(Variable, Variable)> for Graph {
-    type Output = i32;
-    fn index(&self, xy: (Variable, Variable)) -> &i32 {
+    type Output = isize;
+    fn index(&self, xy: (Variable, Variable)) -> &isize {
         self.index((xy.0.id(), xy.1.id()))
     }
 }
 impl IndexMut<(Variable, Variable)> for Graph {
-    fn index_mut(&mut self, xy: (Variable, Variable)) -> &mut i32 {
+    fn index_mut(&mut self, xy: (Variable, Variable)) -> &mut isize {
         self.index_mut((xy.0.id(), xy.1.id()))
     }
 }
 impl Index<(usize, usize)> for Graph {
-    type Output = i32;
-    fn index(&self, xy: (usize, usize)) -> &i32 {
+    type Output = isize;
+    fn index(&self, xy: (usize, usize)) -> &isize {
         let off = self.offset(xy.0, xy.1);
         &self.adj_matrix[off]
     }
 }
 impl IndexMut<(usize, usize)> for Graph {
-    fn index_mut(&mut self, xy: (usize, usize)) -> &mut i32 {
+    fn index_mut(&mut self, xy: (usize, usize)) -> &mut isize {
         let off = self.offset(xy.0, xy.1);
         &mut self.adj_matrix[off]
     }
